Services navigation bar configuration and services viewModel creation
