+++ app/src/main/java/com/example/bankodemia/HomeActivity.kt
@@ -26,7 +26,7 @@ class HomeActivity : AppCompatActivity() {
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications,R.id.cardsFragment
+                R.id.navigation_home, R.id.servicesFragment,R.id.cardsFragment
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
