+++ app/src/main/java/com/example/bankodemia/UI/View/ServicesFragment.kt
@@ -1,0 +1,37 @@
+package com.example.bankodemia.UI.View
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.lifecycle.ViewModelProvider
+import com.example.bankodemia.R
+import com.example.bankodemia.UI.viewModel.CardsViewModel
+import com.example.bankodemia.UI.viewModel.ServicesViewModel
+import com.example.bankodemia.databinding.FragmentCardsBinding
+import com.example.bankodemia.databinding.FragmentServicesBinding
+
+
+class ServicesFragment : Fragment() {
+    private var _binding: FragmentServicesBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val servicesViewModel =
+            ViewModelProvider(this).get(ServicesViewModel::class.java)
+
+        _binding = FragmentServicesBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
