+++ app/src/main/java/com/example/bankodemia/UI/viewModel/ServicesViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.bankodemia.UI.viewModel
+
+import androidx.lifecycle.ViewModel
+
+class ServicesViewModel : ViewModel() {
+}
