+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -12,8 +12,8 @@
         android:title="@string/cards" />
 
     <item
-        android:id="@+id/navigation_notifications"
+        android:id="@+id/servicesFragment"
         android:icon="@drawable/ic_services"
-        android:title="@string/title_notifications" />
+        android:title="@string/servicies_lowercase" />
 
 </menu>
