+++ app/src/main/res/values/strings.xml
@@ -119,4 +119,5 @@
     <string name="helper_fields_no_match">Los campos no coinciden</string>
     <string name="make_transfer">Hacer Transferencia</string>
     <string name="cards">Tarjetas</string>
+    <string name="servicies_lowercase">Servicios</string>
 </resources>
